move all front end functions out of app-api - fixes bug with "can't resolve 'net'"
