+++ __tests__/lib/app-api.test.ts
@@ -1,39 +1,5 @@
-import { alreadyFilledIn, base64EncodeString, isJpegOrPngDataURL } from "@/lib/app-api"
+import { alreadyFilledIn } from "@/lib/app-api"
 import { User } from "@/types/db"
-
-describe("base64EncodeString", () => {
-  it("encodes a string to base64", () => {
-    expect(base64EncodeString("hello world")).toEqual("aGVsbG8gd29ybGQ=")
-    expect(base64EncodeString("foo bar baz")).toEqual("Zm9vIGJhciBiYXo=")
-  })
-})
-
-describe("isJpegOrPngDataURL", () => {
-  it("should return true for a valid JPEG data URL", () => {
-    const dataURL = "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEASABIAAD"
-    expect(isJpegOrPngDataURL(dataURL)).toBe(true)
-  })
-
-  it("should return true for a valid PNG data URL", () => {
-    const dataURL = "data:image/png;base64,iVBORw0KGg"
-    expect(isJpegOrPngDataURL(dataURL)).toBe(true)
-  })
-
-  it("should return false for an invalid data URL", () => {
-    const dataURL = "data:image/gif;base64,R0lGODlhAQ"
-    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
-  })
-
-  it("should return false for a non-base64 data URL", () => {
-    const dataURL = "data:image/jpeg;url=https://example.com/image.jpg"
-    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
-  })
-
-  it("should return false for a non-JPEG/PNG data URL", () => {
-    const dataURL = "data:image/gif;base64,R0lGODlhAQ..."
-    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
-  })
-})
 
 describe("alreadyFilledIn", () => {
   // Mocked document snapshot data
