+++ __tests__/lib/util.test.ts
@@ -1,8 +1,10 @@
 import {
+  base64EncodeString,
   fetchJsonData,
   formatDate,
   formatDateHtml,
   formatDateHtmlReverse,
+  isJpegOrPngDataURL,
   multipleClasses,
   postJsonData,
 } from "@/lib/util"
@@ -150,5 +152,39 @@ describe("fetchJsonData", () => {
     } as unknown as Response)
 
     await expect(fetchJsonData(url, accessToken)).rejects.toThrow()
+  })
+})
+
+describe("base64EncodeString", () => {
+  it("encodes a string to base64", () => {
+    expect(base64EncodeString("hello world")).toEqual("aGVsbG8gd29ybGQ=")
+    expect(base64EncodeString("foo bar baz")).toEqual("Zm9vIGJhciBiYXo=")
+  })
+})
+
+describe("isJpegOrPngDataURL", () => {
+  it("should return true for a valid JPEG data URL", () => {
+    const dataURL = "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEASABIAAD"
+    expect(isJpegOrPngDataURL(dataURL)).toBe(true)
+  })
+
+  it("should return true for a valid PNG data URL", () => {
+    const dataURL = "data:image/png;base64,iVBORw0KGg"
+    expect(isJpegOrPngDataURL(dataURL)).toBe(true)
+  })
+
+  it("should return false for an invalid data URL", () => {
+    const dataURL = "data:image/gif;base64,R0lGODlhAQ"
+    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
+  })
+
+  it("should return false for a non-base64 data URL", () => {
+    const dataURL = "data:image/jpeg;url=https://example.com/image.jpg"
+    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
+  })
+
+  it("should return false for a non-JPEG/PNG data URL", () => {
+    const dataURL = "data:image/gif;base64,R0lGODlhAQ..."
+    expect(isJpegOrPngDataURL(dataURL)).toBe(false)
   })
 })
