+++ src/pages/api/details.ts
@@ -1,12 +1,8 @@
 import { z } from "zod"
 
 import { storageBucket, user } from "@/lib/db"
-import {
-  AuthorisedHanlder,
-  createAuthorisedHandler,
-  isJpegOrPngDataURL,
-  parseRequestBody,
-} from "@/lib/app-api"
+import { AuthorisedHanlder, createAuthorisedHandler, parseRequestBody } from "@/lib/app-api"
+import { isJpegOrPngDataURL } from "@/lib/util"
 
 async function uploadImage(dataUrl: string, path: string): Promise<string> {
   const extension = dataUrl.substring("data:image/".length, dataUrl.indexOf(";base64"))
